usb: musb: gadget: clean up ack handling in ep0_rxstate
Originally, ep0_rxstate always set some csr flags at the end of the
function. If however the completion handler chose to stall the device,
it had to be ensured the csr flags didn't get overwritten nor lost. The
stall function sets all flags in ackpend and then sets ackpend to 0, so
if ackpend was still intact after giveback, we didn't stall and csr is
yet to be set, if it wasn't, stall handler set the csr flags for us and
we can bail out.

There are five cases now:
- req == NULL (data not yet received or no buffer queued) -> instantly
  set csr flags in ep0_rxstates
- Giveback will call into ep0_send_ack for implicit status case
  directly -> set ackpend but never use it, then clear in ep0_send_ack
- Giveback calls into completion handler which stalls instantly ->
  ackpend will be cleared when giveback returns, which used to be the
  only possibility for stalls at that stage, now handled the same as a
  deferred stall
- Giveback calls into completion handler which stalls in the future ->
  stalls can now be deferred and hence we shouldn't clear ackpend that
  early, just keep ackpend there until it is cleared by either stalling
  or by calling ep0_send_ack.
- Gadget driver explicitly queues an empty packet in the future and
  ep0_send_ack is called -> clear ackpend only at that point to avoid
  races.

We also need to make sure ep0_send_ack is actually called when in
STATUSIN stage, because the code to move from SETUP to RX stage
conflicted with the code we wanted to run. That never was noticed
because the code happens to do basically the same thing, however, the
request will never be given back in the SETUP to RX case, which will
leak it. Hence first check if we are in STATUSIN and only if we aren't,
let the SETUP to RX code run.
Also removing some dead code from ep0_send_ack which can only get called
in MUSB_EP0_STAGE_STATUSIN state.

Note: this patch was not tested on musb hardware.

Change-Id: I323c49f2c813e0cb4632012eac4aca00f437bdf1